Added Support for USB3 Cameras. Resolves #14.
Contribtor: Louis-Romain JOLY <https://github.com/lr-sncf>

Squashed commit of the following:

commit 17b0ecec33d7296103ed3e33eee13fe2ef3317ef
Author: Boitumelo Ruf <Boitumelo.Ruf@iosb.fraunhofer.de>
Date:   Tue Nov 5 15:30:08 2024 +0000

    Cherry picked 6e44d35 - Added sercie to manually trigger computation of white balance

commit d20519a6c80ebf2b6d59cd5ff788e18d9d3f88b3
Author: Boitumelo Ruf <Boitumelo.Ruf@iosb.fraunhofer.de>
Date:   Tue Nov 5 15:03:18 2024 +0000

    Updated Readme for USB3Support

commit e0f4dbe4dc60970d57a32b4c95565db2a34ba5b2
Author: Boitumelo Ruf <Boitumelo.Ruf@iosb.fraunhofer.de>
Date:   Tue Oct 29 14:27:12 2024 +0000

    Added usb3support branch to ci config

commit 7abae4108ce89bdab021b114fe4ca39f5d24b406
Author: Boitumelo Ruf <Boitumelo.Ruf@iosb.fraunhofer.de>
Date:   Tue Oct 29 14:23:43 2024 +0000

    Minor code refactoring

commit 661c5677b685ce98c6c44b5b84c861ee3ff25230
Author: Boitumelo Ruf <Boitumelo.Ruf@iosb.fraunhofer.de>
Date:   Tue Oct 29 14:19:06 2024 +0000

    Cherry picked implementation of dynamic parameters (a9ee558ed46a246b8c823fd2710e0a9725e29326)

commit 1eb7bbafdca3fa2f29fd8814c4c88cf6df119895
Author: Boitumelo Ruf <Boitumelo.Ruf@iosb.fraunhofer.de>
Date:   Tue Oct 29 13:50:02 2024 +0000

    Deleted unused config files

commit 88b7dc2161259551c1d62b3709a277736dead85b
Author: Boitumelo Ruf <Boitumelo.Ruf@iosb.fraunhofer.de>
Date:   Tue Oct 29 13:31:31 2024 +0000

    Fixed test failures

commit 2663e16145e134f2d8c531eb1bd5ec1fa82e79aa
Author: Boitumelo Ruf <Boitumelo.Ruf@iosb.fraunhofer.de>
Date:   Tue Oct 29 13:29:35 2024 +0000

    Updated launch files

commit f30eca0e5d4ff440a57b278273b59ee107e8c485
Author: Boitumelo Ruf <Boitumelo.Ruf@iosb.fraunhofer.de>
Date:   Mon Oct 14 13:56:35 2024 +0000

    Refactored code of camera drivers and introduced common base class

commit 93a399991b190466b34f16dd4a4fbbef49862b15
Author: Boitumelo Ruf <Boitumelo.Ruf@iosb.fraunhofer.de>
Date:   Wed Oct 2 10:04:31 2024 +0000

    removed settings.json

commit 441379a979030c10b1cba05508c688ee9298ebcc
Merge: 6e44d35 a498ba6
Author: Boitumelo Ruf <boitumelo.ruf@iosb.fraunhofer.de>
Date:   Wed Oct 2 11:49:38 2024 +0200

    Merge pull request #38 from lr-sncf/usb3support

    added support for GeniCam usb3 cameras, failed check are only py code styles

commit a498ba6827d2db2894f04fd4928a7190dcb35b4a
Author: Louis-Romain JOLY <louis-romain.joly@sncf.fr>
Date:   Tue Oct 1 08:03:44 2024 +0200

    minor correction in comments, add another diagnostic YAML file

commit d68e7dce9b4abce641390df65e8e141236a192e3
Author: Louis-Romain JOLY <louis-romain.joly@sncf.fr>
Date:   Fri Sep 27 10:27:04 2024 +0200

    working version

commit d02f727f6096918f1a3c40bfc9439a829e854250
Author: Louis-Romain JOLY <louis-romain.joly@sncf.fr>
Date:   Wed Sep 25 09:37:36 2024 +0200

    init commit for support
