+++ camera_aravis2/config/camera_diagnostics_example.yaml
@@ -3,16 +3,13 @@
   Selectors:
     - FeatureName: DeviceTemperatureSelector
       Type: string
-      Value: Mainboard    
-    # - FeatureName: DeviceTemperatureSelector
-    #   Type: string
-    #   Value: Sensor
-- FeatureName: PtpEnable
-  Type: bool
-- FeatureName: PtpClockAccuracy
-  Type: string
-- FeatureName: PtpOperationMode
-  Type: string
-- FeatureName: PtpStatus
-  Type: string
+      Value: Sensor    
+#- FeatureName: PtpEnable
+#  Type: bool
+#- FeatureName: PtpClockAccuracy
+#  Type: string
+#- FeatureName: PtpOperationMode
+#  Type: string
+#- FeatureName: PtpStatus
+#  Type: string
