+++ camera_aravis2/config/camera_info_example_gv.yaml
