+++ camera_aravis2/config/camera_info_example_uv.yaml
@@ -1,0 +1,26 @@
+image_width: 1920
+image_height: 1200
+camera_name: camera_uv_vis
+camera_matrix:
+  rows: 3
+  cols: 3
+  data: [1778.591345, 0.000000, 964.276439,
+         0.000000, 1789.714953, 804.364652,
+         0.000000, 0.000000, 1.000000]
+distortion_model: plumb_bob
+distortion_coefficients:
+  rows: 1
+  cols: 5
+  data: [-0.149383, 0.142808, 0.002512, -0.003641, 0.000000]
+rectification_matrix:
+  rows: 3
+  cols: 3
+  data: [ 0.969314, -0.024000, 0.244652,
+          0.025128, 0.999683, -0.001489,
+          -0.244539, 0.007591, 0.969610]
+projection_matrix:
+  rows: 3
+  cols: 4
+  data: [ 2720.246602, 0.000000, 710.195835, 0.000000,
+          0.000000, 2720.246602, 798.603073, 0.000000,
+          0.000000, 0.000000, 1.000000, 0.000000]
