+++ camera_aravis2/include/camera_aravis2/camera_aravis_node_base.h
@@ -99,7 +99,7 @@ class CameraAravisNodeBase : public rclcpp::Node
     /**
      * @brief Set the up launch parameters.
      */
-    virtual void setUpParameters();
+    virtual void setupParameters();
 
     /**
      * @brief Discover attached camera devices found by Aravis and open device specified by guid.
