+++ camera_aravis2/include/camera_aravis2/camera_driver_uv.h
@@ -28,16 +28,38 @@
 
 #ifndef CAMERA_ARAVIS2__CAMERA_DRIVER_UV_H_
 #define CAMERA_ARAVIS2__CAMERA_DRIVER_UV_H_
+// Yaml-cpp
+#include <yaml-cpp/yaml.h>
+
+// Std
+#include <memory>
+#include <string>
+#include <unordered_map>
+#include <utility>
+#include <vector>
+
+// Aravis
+extern "C"
+{
+#include "aravis-0.8/arv.h"
+}
 
 // ROS
+#include <camera_info_manager/camera_info_manager.hpp>
+#include <image_transport/image_transport.hpp>
 #include <rclcpp/rclcpp.hpp>
 
 // camera_aravis2
-#include "camera_aravis2/camera_aravis_node_base.h"
+#include "camera_aravis2/camera_driver.h"
+#include "camera_aravis2/concurrent_queue.hpp"
+#include "camera_aravis2/config_structs.h"
+#include "camera_aravis2/conversion_utils.h"
+#include "camera_aravis2/image_buffer_pool.h"
+#include <camera_aravis2_msgs/msg/camera_diagnostics.hpp>
 
 namespace camera_aravis2
 {
-class CameraDriverUv : public CameraAravisNodeBase
+class CameraDriverUv : public CameraDriver
 {
     //--- METHOD DECLARATION ---//
 
@@ -48,7 +70,7 @@ class CameraDriverUv : public CameraAravisNodeBase
      * @param[in] options Node options.
      */
     explicit CameraDriverUv(const rclcpp::NodeOptions& options =
-                              rclcpp::NodeOptions().allow_undeclared_parameters(true));
+                              rclcpp::NodeOptions());
 
     /**
      * @brief Default destructor.
@@ -57,12 +79,46 @@ class CameraDriverUv : public CameraAravisNodeBase
     virtual ~CameraDriverUv();
 
   protected:
-    //--- FUNCTION DECLARATION ---//
+    /**
+     * @brief Method to set UsbVision specific transport layer control settings
+     * of the camera.
+     *
+     * @return True if successful. False, otherwise.
+     */
+    [[nodiscard]] bool setTechSpecificTlControlSettings() override;
 
-  protected:
+    /**
+     * @brief Discover number of available camera streams. In case of USB3 cameras this will
+     * return 1.
+     *
+     */
+    int discoverNumberOfStreams() override;
+
+    /**
+     * @brief Tune specific parameters for Aravis streams.
+     *
+     * @param[in] p_stream Pointer to Aravis stream.
+     */
+    void tuneArvStream(ArvStream* p_stream) const override;
+
+    /**
+     * @brief Callback method to inject short processing routines after the
+     * publication of each frame.
+     *
+     * This is called within the processing loop of processStreamBuffer.
+     *
+     * @note This should not hold heavy processing as it will the delay the processing of the next
+     * frame within the stream buffer.
+     *
+     * @param[in] stream_id ID of stream for which the buffer is to be processed.
+     */
+    void postFrameProcessingCallback(const uint stream_id) override;
+
     //--- MEMBER DECLARATION ---//
 
   protected:
+    /// Pointer to Uv-Specific transport layer control settings.
+    std::shared_ptr<UvTransportLayerControl> p_uv_tl_control_;
 };
 
 }  // namespace camera_aravis2
