+++ camera_aravis2/launch/camera_driver_uv_example.launch.py
@@ -43,60 +43,38 @@ os.environ['RCUTILS_CONSOLE_OUTPUT_FORMAT'] = '{time}: [{name}] [{severity}]\t{m
 def generate_launch_description():
 
     example_package_node = Node(
-        name='camera_driver_gv_example',
+        name='camera_driver_uv_example',
         package='camera_aravis2',
-        executable='camera_driver_gv',
+        executable='camera_driver_uv',
         output='screen',
         emulate_tty=True,
         # arguments=['--ros-args', '--log-level', 'debug'],
         parameters=[
                 {
                     # Driver-specific parameters
-                    'guid': 'Allied Vision-Alvium G1-240c-05P3C',
-                    'frame_id': 'camera_gv',
+                    'guid': 'IDS Imaging Development Systems GmbH-1409f4a3c131-4104372529',
+                    'frame_id': 'camera_uv',
                     'stream_names': ['vis'],
                     'camera_info_urls': ['file://' + os.path.join(
                         get_package_share_directory('camera_aravis2'),
-                        'config/camera_info_example_gv.yaml')],
+                        'config/camera_info_example_uv.yaml')],
                     'dynamic_parameters_yaml_url': os.path.join(
                         get_package_share_directory('camera_aravis2'),
                         'config/dynamic_parameters_example.yaml'),
-                    'diagnostic_yaml_url': os.path.join(
-                        get_package_share_directory('camera_aravis2'),
-                        'config/camera_diagnostics_example.yaml'),
                     'verbose': False,
 
                     # GenICam-specific parameters
                     'DeviceControl': {
                         'DeviceLinkThroughputLimit': 125000000,
-                        'DeviceLinkThroughputLimitMode': 'On'
-                    },
-                    'TransportLayerControl': {
-                        'GevSCPSPacketSize': 9000,
-                        'PtpEnable': True
                     },
                     'ImageFormatControl': {
-                        'BEGIN': {
-                            'BinningSelector': 'Digital'
-                        },
-                        'PixelFormat': ['BayerRG8', 'Mono8'],
+                        'PixelFormat': ['BayerGR8'],
                         'Width': 1920,
                         'Height': 1200
                     },
                     'AcquisitionControl': {
-                        'ExposureTime': 10000.0,
-                        'ExposureAuto': 'Off',
-                        'ExposureMode': 'Timed',
                         'AcquisitionFrameRateEnable': True,
                         'AcquisitionFrameRate': 30.0
-                    },
-                    'AnalogControl': {
-                        'GainAuto': 'Continuous',
-                        'BalanceWhiteAuto': 'Off',
-                        'BalanceRatio': {
-                            'Red': 1.6,
-                            'Blue': 2.0
-                        }
                     }
                 }
             ]
