+++ camera_aravis2/src/camera_driver_uv.cpp
@@ -28,23 +28,177 @@
 
 #include "camera_aravis2/camera_driver_uv.h"
 
+// Std
+#include <chrono>
+#include <iostream>
+#include <thread>
+
+// ROS
+#include <rclcpp/time.hpp>
+
+// camera_aravis2
+#include "camera_aravis2/conversion_utils.h"
+#include "camera_aravis2/error.h"
+
 namespace camera_aravis2
 {
 
 //==================================================================================================
 CameraDriverUv::CameraDriverUv(const rclcpp::NodeOptions& options) :
-  CameraAravisNodeBase("camera_driver_uv", options)
+  CameraDriver("camera_driver_uv", options)
 {
-    RCLCPP_FATAL(logger_, "USB3 Devices are currently not supported.");
-    RCLCPP_FATAL(logger_,
-                 "Help Wanted: https://github.com/FraunhoferIOSB/camera_aravis2/issues/14");
-    return;
+    //--- setup parameters
+    setupParameters();
+
+    is_verbose_enable_ = get_parameter("verbose").as_bool();
+
+    //--- get parameter overrides, i.e. all parameters, including those that are not declared
+    parameter_overrides_ = this->get_node_parameters_interface()->get_parameter_overrides();
+
+    //--- open camera device
+    ASSERT_SUCCESS(discoverAndOpenCameraDevice());
+
+    //--- check if UV Device.
+    if (!arv_camera_is_uv_device(p_camera_))
+    {
+        RCLCPP_FATAL(logger_, "Camera is no USB3-Vision Device.");
+        return;
+    }
+
+    //--- set up structs holding relevant information of camera streams
+    ASSERT_SUCCESS(setupCameraStreamStructs());
+
+    //--- set device control settings
+    ASSERT_SUCCESS(setDeviceControlSettings());
+
+    //--- set transport layer control settings
+    p_uv_tl_control_ = std::make_shared<UvTransportLayerControl>();
+    p_tl_control_    = std::dynamic_pointer_cast<GenTransportLayerControl>(p_uv_tl_control_);
+
+    //--- set image format control settings
+    ASSERT_SUCCESS(setImageFormatControlSettings());
+
+    //--- set acquisition control settings
+    ASSERT_SUCCESS(setAcquisitionControlSettings());
+
+    //--- set analog control settings
+    ASSERT_SUCCESS(setAnalogControlSettings());
+
+    //--- initialize services
+    ASSERT_SUCCESS(initializeServices());
+
+    //--- load dynamic parameters
+    setupDynamicParameters();
+
+    //--- load diagnostics
+    setupCameraDiagnosticPublisher();
+
+    //--- print currently applied camera configuration
+    printCameraConfiguration();
+
+    //--- spawn camera stream in thread, so that initialization is not blocked
+    is_spawning_         = true;
+    spawn_stream_thread_ = std::thread(&CameraDriverUv::spawnCameraStreams, this);
 }
 
 //==================================================================================================
 CameraDriverUv::~CameraDriverUv()
 {
+    // Guarded error object
+    GuardedGError err;
+
+    //--- stop acquisition
+    if (p_device_)
+    {
+        arv_device_execute_command(p_device_, "AcquisitionStop", err.ref());
+        CHECK_GERROR_MSG(err, logger_, "In executing 'AcquisitionStop'.");
+    }
+
+    //--- stop emitting signals for streams
+    for (uint i = 0; i < streams_.size(); i++)
+        if (streams_[i].p_arv_stream)
+            arv_stream_set_emit_signals(streams_[i].p_arv_stream, FALSE);
+
+    //--- join spawning thread
+    is_spawning_ = false;
+    if (spawn_stream_thread_.joinable())
+        spawn_stream_thread_.join();
+
+    //--- join diagnostic thread
+    is_diagnostics_published_ = false;
+    if (diagnostic_thread_.joinable())
+    {
+        diagnostic_thread_.join();
+    }
+
+    //--- join buffer threads
+    for (uint i = 0; i < streams_.size(); i++)
+    {
+        Stream& stream = streams_[i];
+
+        stream.is_buffer_processed = false;
+
+        //--- push empty object into queue to do final loop
+        stream.buffer_queue.push(std::make_pair(nullptr, nullptr));
+
+        if (stream.buffer_processing_thread.joinable())
+            stream.buffer_processing_thread.join();
+    }
+
+    //--- print stream statistics
+    printStreamStatistics();
+
+    //--- unref pointers
+    for (uint i = 0; i < streams_.size(); i++)
+    {
+        if (streams_[i].p_arv_stream)
+            g_object_unref(streams_[i].p_arv_stream);
+    }
+
     //--- camera / device object is unreferences in parent class
+}
+
+//==================================================================================================
+[[nodiscard]] bool CameraDriverUv::setTechSpecificTlControlSettings()
+{
+    std::shared_ptr<UvTransportLayerControl> p_uv_tl_control =
+      std::dynamic_pointer_cast<UvTransportLayerControl>(p_tl_control_);
+
+    //--- check if transport layer control is correctly initialized
+    if (!p_uv_tl_control_)
+    {
+        RCLCPP_FATAL(logger_, "%s:  Pointer to UvTransportLayerControl is not initialized. ",
+                     __PRETTY_FUNCTION__);
+
+        return false;
+    }
+
+    return true;
+}
+
+//==================================================================================================
+int CameraDriverUv::discoverNumberOfStreams()
+{
+    return 1;
+}
+
+//==================================================================================================
+void CameraDriverUv::tuneArvStream(ArvStream* p_stream) const
+{
+    if (!p_stream)
+        return;
+
+    if (!ARV_IS_UV_STREAM(p_stream))
+    {
+        RCLCPP_ERROR(logger_, "Stream is not a UV_STREAM");
+        return;
+    }
+}
+
+//==================================================================================================
+void CameraDriverUv::postFrameProcessingCallback(const uint stream_id)
+{
+    RCL_UNUSED(stream_id);
 }
 
 }  // end namespace camera_aravis2
